Create `typeDefs` file with all Gatsby schema customization
